requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
pyAesCrypt==6.0.0
pycparser==2.21
six==1.16.0
Werkzeug==2.2.2
